+++ kubernetes-model-generator/kubernetes-model-flowcontrol/cmd/generate/generate.go
@@ -20,8 +20,10 @@ import (
   "encoding/json"
   "fmt"
 
+  flowcontrolv1alpha1 "k8s.io/api/flowcontrol/v1alpha1"
   flowcontrolv1beta1 "k8s.io/api/flowcontrol/v1beta1"
   flowcontrolv1beta2 "k8s.io/api/flowcontrol/v1beta2"
+  flowcontrolv1beta3 "k8s.io/api/flowcontrol/v1beta3"
 
 
   "log"
@@ -35,6 +37,10 @@ import (
 )
 
 type Schema struct {
+  V1alpha1FlowSchema                       flowcontrolv1alpha1.FlowSchema
+  V1alpha1FlowSchemaList                   flowcontrolv1alpha1.FlowSchemaList
+  V1alpha1PriorityLevelConfiguration       flowcontrolv1alpha1.PriorityLevelConfiguration
+  V1alpha1PriorityLevelConfigurationList   flowcontrolv1alpha1.PriorityLevelConfigurationList
   V1beta1FlowSchema                        flowcontrolv1beta1.FlowSchema
   V1beta1FlowSchemaList                    flowcontrolv1beta1.FlowSchemaList
   V1beta1PriorityLevelConfiguration        flowcontrolv1beta1.PriorityLevelConfiguration
@@ -43,6 +49,10 @@ type Schema struct {
   V1beta2FlowSchemaList                    flowcontrolv1beta2.FlowSchemaList
   V1beta2PriorityLevelConfiguration        flowcontrolv1beta2.PriorityLevelConfiguration
   V1beta2PriorityLevelConfigurationList    flowcontrolv1beta2.PriorityLevelConfigurationList
+  V1beta3FlowSchema                        flowcontrolv1beta3.FlowSchema
+  V1beta3FlowSchemaList                    flowcontrolv1beta3.FlowSchemaList
+  V1beta3PriorityLevelConfiguration        flowcontrolv1beta3.PriorityLevelConfiguration
+  V1beta3PriorityLevelConfigurationList    flowcontrolv1beta3.PriorityLevelConfigurationList
 }
 
 func main() {
@@ -52,8 +62,10 @@ func main() {
     {"k8s.io/apimachinery/pkg/version", "", "io.fabric8.kubernetes.api.model.version", "kubernetes_apimachinery_pkg_version_", false},
     {"k8s.io/apimachinery/pkg/apis/meta/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_apimachinery_", false},
     {"k8s.io/api/core/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_core_", false},
+    {"k8s.io/api/flowcontrol/v1alpha1", "flowcontrol.apiserver.k8s.io", "io.fabric8.kubernetes.api.model.flowcontrol.v1alpha1", "kubernetes_flowcontrol_v1alpha1_", true},
     {"k8s.io/api/flowcontrol/v1beta1", "flowcontrol.apiserver.k8s.io", "io.fabric8.kubernetes.api.model.flowcontrol.v1beta1", "kubernetes_flowcontrol_v1beta1_", true},
     {"k8s.io/api/flowcontrol/v1beta2", "flowcontrol.apiserver.k8s.io", "io.fabric8.kubernetes.api.model.flowcontrol.v1beta2", "kubernetes_flowcontrol_v1beta2_", true},
+    {"k8s.io/api/flowcontrol/v1beta3", "flowcontrol.apiserver.k8s.io", "io.fabric8.kubernetes.api.model.flowcontrol.v1beta3", "kubernetes_flowcontrol_v1beta3_", true},
   }
 
   typeMap := map[reflect.Type]reflect.Type{